fix: some sites remove our css file (#377)
* fix: some sites remove our css file

it looks timing related - as it doesn't happen if we load a little later. Likely React wiping the dom of changes

This change adds a mutation observer to the parent of the css file. then removes it after 10s

only allow replace to run once.
logging if hit maximum readds

AP-2244